requirements.txt
aiogram==3.9.0
asyncpg
APScheduler
python-decouple
aiohttp
gspread
openpyxl
matplotlib
pandas